Fix status code being passed as database ID (#35)
In `notion_search_id` we are returning one of three values: `database_id`, `None`, and `status_code`. We don't use the `status_code` for branching but pass the value from the function assuming it is the `database_id` to the next functions. This results in us querying Notion for a `database_id` value of the `status_code` which is incorrect